requirements.txt
fastapi==0.95.1
uvicorn==0.34.0
sqlalchemy==1.4.44
pydantic==2.10.6
jinja2==3.1.5
openai==1.63.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
alembic== 1.14.1
asyncpg==0.27.0
psycopg2-binary==2.9.6


